Refactor HammingOneGenerator and related classes to improve memory allocation flags and variable types; update file handling to use size_t for loop indices.

+++ src/intvlk/PerFrameData.hpp
@@ -43,7 +43,7 @@ namespace intvlk
         {
             std::vector<PerFrameData> perFrameData{};
             perFrameData.reserve(queuedFramesCount);
-            for (uint32_t i{0}; i < queuedFramesCount; ++i)
+            for (size_t i{0}; i < queuedFramesCount; ++i)
             {
                 perFrameData.emplace_back(device, queueFamilyIndex);
             }
